fix bug with wrong share

+++ internal/domain/chart/chart.go
@@ -3,9 +3,9 @@ package chart
 const DateFormat = "2006-01-02"
 
 type Candle struct {
-	Time  int `json:"time"`
-	Open  int `json:"open"`
-	Close int `json:"close"`
-	High  int `json:"high"`
-	Low   int `json:"low"`
+	Time  int64   `json:"time"`
+	Open  float64 `json:"open"`
+	Close float64 `json:"close"`
+	High  float64 `json:"high"`
+	Low   float64 `json:"low"`
 }
